Courses & Lectures midway

+++ database/migrations/2020_12_07_210708_create_lectures_table.php
@@ -19,7 +19,7 @@ class CreateLecturesTable extends Migration
             $table->text('content');
             $table->boolean('containsVideo');
             $table->text('urlVideo')->nullable();
-            $table->integer('duration');
+            $table->unsignedSmallInteger('duration');
             $table->foreignId('course_id')->constrained()->cascadeOnDelete();
             $table->timestamps();
         });
